small frontend refactoring
