+++ client/src/pages/AuthStyles.js
